# [COST] Improve `NetworkCost` Performance