@chia7712 2023-03-04
另外這邊用到 Cache 似乎有點大材小用，要否用 Map 取代就好？